+++ home/models.py
@@ -167,6 +167,21 @@ class VariationBlock(blocks.StructBlock):
     )
 
 
+class NextMessageButton(blocks.StructBlock):
+    title = blocks.CharBlock(
+        help_text="text for the button, up to 20 characters.",
+        validators=(MaxLengthValidator(20),),
+    )
+
+
+class GoToPageButton(blocks.StructBlock):
+    title = blocks.CharBlock(
+        help_text="text for the button, up to 20 characters.",
+        validators=(MaxLengthValidator(20),),
+    )
+    page = blocks.PageChooserBlock(help_text="page the button should go to")
+
+
 class WhatsappBlock(blocks.StructBlock):
     MEDIA_CAPTION_MAX_LENGTH = 1024
 
@@ -179,10 +194,16 @@ class WhatsappBlock(blocks.StructBlock):
         validators=(MaxLengthValidator(4096),),
     )
     variation_messages = blocks.ListBlock(VariationBlock(), default=[])
+    # TODO: next_prompt is deprecated, and should be removed in the next major version
     next_prompt = blocks.CharBlock(
         help_text="prompt text for next message",
         required=False,
         validators=(MaxLengthValidator(20),),
+    )
+    buttons = blocks.StreamBlock(
+        [("next_message", NextMessageButton()), ("go_to_page", GoToPageButton())],
+        required=False,
+        max_num=3,
     )
 
     class Meta: