+++ home/tests/test_api.py
@@ -12,6 +12,7 @@ from .utils import create_page
 
 from home.models import (  # isort:skip
     ContentPage,
+    HomePage,
     OrderedContentSet,
     PageView,
     VariationBlock,
@@ -329,6 +330,42 @@ class PaginationTestCase(TestCase):
 
         self.assertEqual(content, {"page": ["Page matching query does not exist."]})
         self.assertEqual(content.get("page"), ["Page matching query does not exist."])
+
+
+class WhatsAppMessagesTestCase(TestCase):
+    def setUp(self):
+        self.user_credentials = {"username": "test", "password": "test"}
+        self.user = get_user_model().objects.create_user(**self.user_credentials)
+        self.client.login(**self.user_credentials)
+
+    def test_whatsapp_detail_view_with_button(self):
+        page = ContentPage(
+            title="test",
+            slug="text",
+            enable_whatsapp=True,
+            whatsapp_body=[
+                {
+                    "type": "Whatsapp_Message",
+                    "value": {
+                        "message": "test message",
+                        "buttons": [
+                            {"type": "next_message", "value": {"title": "Tell me more"}}
+                        ],
+                    },
+                }
+            ],
+        )
+        homepage = HomePage.objects.first()
+        homepage.add_child(instance=page)
+        page.save_revision().publish()
+
+        response = self.client.get(f"/api/v2/pages/{page.id}/?whatsapp=true&message=1")
+        content = response.json()
+        [button] = content["body"]["text"]["value"]["buttons"]
+        button.pop("id")
+        self.assertEqual(
+            button, {"type": "next_message", "value": {"title": "Tell me more"}}
+        )
 
 
 class OrderedContentSetTestCase(TestCase):